+++ backend/app/models/goal.py
@@ -1,5 +1,5 @@
 from datetime import date, datetime
-from sqlalchemy import Integer, Float, Date, ForeignKey, Text, String, UniqueConstraint
+from sqlalchemy import BigInteger, Integer, Float, Date, ForeignKey, Text, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.database import Base
 
@@ -9,7 +9,7 @@ class Goal(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     name: Mapped[str] = mapped_column(String(255), nullable=False)
-    target_amount_inr: Mapped[int] = mapped_column(Integer, nullable=False)  # paise
+    target_amount_inr: Mapped[int] = mapped_column(BigInteger, nullable=False)  # paise
     target_date: Mapped[date] = mapped_column(Date, nullable=False)
     assumed_return_pct: Mapped[float] = mapped_column(Float, default=12.0, nullable=False)
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
